+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:atmos/atmos.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -14,53 +15,45 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
       ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const HomePage(),
     );
   }
 }
 
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
+final counterAtom = atom(0);
+final doubledAtom = computed((get) => get(counterAtom) * 2);
 
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-
-        title: Text(widget.title),
-      ),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text('You have pushed the button this many times:'),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
+        child: AtomBuilder(
+          atom: counterAtom,
+          builder: (context, value, setValue) {
+            return Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text('Count: $value', style: TextStyle(fontSize: 40)),
+                AtomBuilder(
+                  atom: doubledAtom,
+                  builder: (context, doubleValue, _) {
+                    return Text(
+                      'Double Count: $doubleValue',
+                      style: TextStyle(fontSize: 40),
+                    );
+                  },
+                ),
+                ElevatedButton(
+                  onPressed: () => setValue(value + 1),
+                  child: Text("Increment"),
+                ),
+              ],
+            );
+          },
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
       ),
     );
   }