+++ lib/atmos.dart
@@ -1,8 +1,6 @@
 /// Minimal Jotai-inspired state management for Flutter with atoms.
 library;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/atom.dart';
+export 'src/atom_builder.dart';
+export 'src/atmos_store.dart';